doc: notes on tooling: circomkit, circomspect, circomscribe, picus, circom-mutator (#1)
add notes on usage and context for the following circom analysis tools in readme:

circomspect
circomscribe
picus
circom --inspect
circomkit

    circom-mutator

This PR adds notes and demos in the readme on 3 tools worth using. picus I was unable to build, circomscribe seemed to have no obvious use case, and circom-mutator seemed like it would take a day before I understood what was going on. The first and the last could be useful, but seemed too high upfront learning curve to figure out in a day (or get to build at all in the case of picus, though Matt says he was able to build it).

I'm going to follow this PR up tomorrow with more notes from a few of the resources I mention at the bottom of this readme, which I didn't get to here.